requirements.txt
opencv-python~=4.10.0.82
streamlit~=1.32.2
numpy~=1.26.4
lxml~=5.1.0
Pillow~=10.1.0
streamlit_image_zoom
tensorflow